Make desugaring an explicit feature

+++ dex-members-list/src/test/testdata/build.sh
@@ -16,13 +16,17 @@ for folder in `find "$DIR" -type d -mindepth 1 -maxdepth 1`; do
   rm -f *.class
   javac *.java
   cp *.class "../../resources/"
-  zip -r "$folder.jar" *.class
+
+  if [ "desugaring" != "$name" ]; then
+    zip -r "$folder.jar" *.class
+    cp "$folder.jar" "../../resources/"
+    dx --dex --output="$folder.dex" "$folder.jar"
+    rm "$folder.jar"
+    cp "$folder.dex" "../../resources/"
+    rm "$folder.dex"
+  fi
+
   rm *.class
-  cp "$folder.jar" "../../resources/"
-  dx --dex --output="$folder.dex" "$folder.jar"
-  rm "$folder.jar"
-  cp "$folder.dex" "../../resources/"
-  rm "$folder.dex"
   popd
 done
